Start passage with chapter:verse number
The chapter and verse numbers are a good place to start the passage to
get easy capatilization, and easy start of sentence sounding things. So
in this change, I added a random chooser of chapter:verse pairs to start
the markov chaining process. I explored the idea in the notebook, then
implmented it in the module.